requirements.txt
aiogram>=3.0.0
aiohttp
python-dotenv
fastapi
uvicorn
pydantic
beautifulsoup4==4.14.3
